Refactoring. Fix the tests.
