+++ src/client/sandbox/event/drag-and-drop/data-store-mode.js
@@ -1,0 +1,6 @@
+// https://html.spec.whatwg.org/multipage/interaction.html#drag-data-store-mode
+export default {
+    readwrite: 'readwrite',
+    readonly:  'readonly',
+    protected: 'protected'
+};
