+++ src/client/sandbox/event/drag-and-drop/data-transfer-item-kind.js
@@ -1,0 +1,5 @@
+// https://html.spec.whatwg.org/multipage/interaction.html#the-drag-data-item-kind
+export default {
+    string: 'string',
+    file:   'file'
+};
